Added the "guess" command

+++ logic/LP.h
@@ -13,9 +13,9 @@
 void guess_solution(Board *board, States *states, double threshold);
 
 /* Returns an array of scores that represents the possible values and their
- * score for the given cell (size of the array is always the board dimension).
+ * score for the given cell (the size of the array is always the board dimension).
  * For example, the array [0, 0.1, 0.33, 0, 0.2, ...] means that the value 2 has a
  * score of 0.1, and the value 3 has a score of 0.33, and so on. */
-int* get_cell_guesses(const Board *board, int row, int column);
+double* get_cell_guesses(const Board *board, int row, int column);
 
 #endif
